refactor: moved the setup functions to the handlers module
