+++ main.py
@@ -1,28 +1,7 @@
 """Main module"""
-from telegram import BotCommand
 from telegram.ext import Updater
 from modules.data import Config
-from modules.handlers import add_handlers, add_jobs
-
-
-def add_commands(updater: Updater):
-    """Adds the list of commands with their description to the bot
-
-    Args:
-        updater: supplied Updater
-    """
-    commands = [
-        BotCommand("start", "presentazione iniziale del bot"),
-        BotCommand("spot", "inizia a spottare"),
-        BotCommand("cancel ",
-                   "annulla la procedura in corso e cancella l'ultimo spot inviato, se non è ancora stato pubblicato"),
-        BotCommand("help ", "funzionamento e scopo del bot"),
-        BotCommand("report", "segnala un utente"),
-        BotCommand("rules ", "regole da tenere a mente"),
-        BotCommand("stats", "visualizza statistiche sugli spot"),
-        BotCommand("settings", "cambia le impostazioni di privacy")
-    ]
-    updater.bot.set_my_commands(commands=commands)
+from modules.handlers import add_commands, add_handlers, add_jobs
 
 
 def main():
